Creation of an Entity withing an EntityManager works.

+++ Frosty/src/Frosty/RenderEngine/RenderEngine.cpp
@@ -361,7 +361,7 @@ namespace Frosty
 	void RenderEngine::RenderTestTriangle()
 	{
 
-		m_Transform.setRotate(glm::vec3(0.0f, m_Rotation += 0.1, 0.0f));//Temp
+		m_Transform.setRotate(glm::vec3(0.0f, m_Rotation += 0.1f, 0.0f));//Temp
 
 		glBindFramebuffer(GL_FRAMEBUFFER, 0);
 		glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
@@ -399,7 +399,7 @@ namespace Frosty
 		lights.emplace_back(l);
 		lights.emplace_back(l2);
 
-		glUniform1ui(glGetUniformLocation(m_ShaderProgramVector[TEST_LIGHT], "nrOfPointLights"), lights.size());
+		glUniform1ui(glGetUniformLocation(m_ShaderProgramVector[TEST_LIGHT], "nrOfPointLights"), (GLuint)lights.size());
 		for (int i = 0; i < lights.size(); i++)
 		{
 
@@ -416,7 +416,7 @@ namespace Frosty
 		lights.emplace_back(l3);
 		lights.emplace_back(l4);
 
-		glUniform1ui(glGetUniformLocation(m_ShaderProgramVector[TEST_LIGHT], "nrOfDirLights"), lights.size());
+		glUniform1ui(glGetUniformLocation(m_ShaderProgramVector[TEST_LIGHT], "nrOfDirLights"), (GLuint)lights.size());
 		for (int i = 0; i < lights.size(); i++)
 		{
 			glUniform4fv(glGetUniformLocation(m_ShaderProgramVector[TEST_LIGHT], ("dirLights[" + std::to_string(i) + "].color").c_str()), 1, &lights[i].color[0]);
